Added  Car Prices Assignment To the repo
The Data and the python code for the Car price assignment:
Car Prices Assignment- Linear Regression
Problem Statement:
Geely Auto aspires to enter the US market by setting up their manufacturing unit there and producing cars locally to give competition to their US and European counterparts.

Which variables are significant in predicting the price of a car? How well those variables describe the price of a car? Model the price of cars with the available independent variables.

Determine All driving factors for the prices of cars in US market for chinese company
Create a Linear Regression Model to predict prices